Allow the use of custom terminal
Currently workon does not allow to use a custom terminal emulator since
it needs to know the option to use to set the working directory.
However the working directory is already set when starting the process.

Update workon to allow the user to specify a custom terminal, using the
-terminal flag.
Currently the -terminal flag is required, so it is not really a flag.