Interpret the command line argument as a module path
Update the main function to interpret the command line argument as a
module path, instead of a directory.  An absolute directory is still
supported but must be inside $GOPATH.

Add the internal/mod package, with the exported Resolve function, to
resolve a pattern to a Module.  Move the isDir function from main.go to
mod.go.

The i3workon module now requires github.com/perillo/gocmd v0.1.0.