Intepret the command line argument as a pattern
Update the internal/mod package to interpret the command line argument
as a pattern.  The pattern has the same syntax as in the go tool, with
the exception that absolute and relative directories are not supported.

Add the MatchModules function, to return all the modules in $GOPATH
matching the pattern.  The function name and implementation as been
adapted from src/cmd/go/internal/search/search.go in the Go source
distribution.