search: don't invoke go list -m
Don't invoke the go list -m command, in the load function; its
performance is not predictable since it may try to access the network.

Update the load function to use golang.org/x/mod/modfile to parse the
go.mod file and synthesize a module.

Simplify the code to check if a module path is inside $GOPATH.  We know
the root directory, the current directory path and the module path, so
the check is very simple.