Decouple module search and load
Decouple the module search, with pattern matching, and module loading.
This is what the go tool does with packages.

In the search package, redefine the Module type to only have fields that
can be deduced based on the directory with an existing go.mod file.
Add the Root field, containing the Go path dir containing the module.

Rename the load function to mkmod and remove the isDir function.

Add a new internal/load package, responsible for loading modules.
The API has been adapted from src/cmd/go/internal/load/pkg.go in the Go
source distribution.

Remove the search/Resolve function and add a resolve function in the
main package that calls the load.Modules function.